requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
aiosmtplib==2.0.2
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
audioread==3.0.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-mail==1.4.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
gradio==4.12.0
gradio_client==0.8.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llamaapi==0.1.36
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
openai==1.6.1
orjson==3.9.10
packaging==23.2
pandas==2.1.4
passlib==1.7.4
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.21.12
pyasn1==0.5.1
PyAudio==0.2.13
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.22
starlette==0.27.0
threadpoolctl==3.2.0
tomlkit==0.12.0
toolz==0.12.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
websockets==11.0.3
yarl==1.9.4
